requirements.txt
praw
vaderSentiment
python-dotenv
nltk